+++ modules/App/assets/vue-components/fields/field-set.js
@@ -26,7 +26,7 @@ export default {
                 output = App.utils.$interpolate(field.opts.display, { /* deprecated */ value, data:value});
             } catch(e) {}
 
-            return output;
+            return DOMPurify.sanitize(output);
         }
     },